requirements.txt
pydantic>=2.0,<3.0
fastapi==0.115.6
uvicorn==0.34.0
openai==1.58.1
httpx==0.28.1
pytest==8.3.4
python-dotenv==1.0.1
pyyaml==6.0.2
openenv-core==0.2.3
